pyproject.toml
[tool.poetry]
name = "data_dynamos_exercise"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.16"
dbt-core = "^1.5.2"
dbt-snowflake = "^1.5.2"
streamlit = "^1.24.1"
plotly = "^5.15.0"
snowflake-snowpark-python = {extras = ["pandas"], version = "^1.5.1"}
watchdog = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
